+++ core/frontend/src/ViewContext.ts
@@ -21,7 +21,7 @@ import { GraphicBranch, GraphicBranchOptions } from "./render/GraphicBranch";
 import { GraphicBuilder, ViewportGraphicBuilderOptions } from "./render/GraphicBuilder";
 import { GraphicList, RenderGraphic } from "./render/RenderGraphic";
 import { RenderPlanarClassifier } from "./internal/render/RenderPlanarClassifier";
-import { RenderSystem, RenderTextureDrape } from "./render/RenderSystem";
+import { RenderSystem, } from "./render/RenderSystem";
 import { RenderTarget } from "./render/RenderTarget";
 import { Scene } from "./render/Scene";
 import { SpatialClassifierTileTreeReference, Tile, TileGraphicType, TileLoadStatus, TileTreeReference } from "./tile/internal";
@@ -29,6 +29,7 @@ import { ViewingSpace } from "./ViewingSpace";
 import { ELEMENT_MARKED_FOR_REMOVAL, ScreenViewport, Viewport, ViewportDecorator } from "./Viewport";
 import { ActiveSpatialClassifier } from "./SpatialClassifiersState";
 import { GraphicType } from "./common/render/GraphicType";
+import { RenderTextureDrape } from "./internal/render/RenderTextureDrape";
 
 /** Provides context for producing [[RenderGraphic]]s for drawing within a [[Viewport]].
  * @public
