+++ core/frontend/src/render/GraphicBranch.ts
@@ -16,7 +16,7 @@ import { RenderClipVolume } from "./RenderClipVolume";
 import { RenderGraphic } from "./RenderGraphic";
 import { RenderMemory } from "./RenderMemory";
 import { RenderPlanarClassifier } from "../internal/render/RenderPlanarClassifier";
-import { RenderTextureDrape } from "./RenderSystem";
+import { RenderTextureDrape } from "../internal/render/RenderTextureDrape";
 import { Range3d, Transform } from "@itwin/core-geometry";
 import { AnimationNodeId } from "../common/internal/render/AnimationNodeId";
 import { GraphicBranchFrustum } from "../internal/render/GraphicBranchFrustum";
